Fancy menu merger (#774)
* merge menu into fancy-menu

* extract buildFilters and chooseMeal out of fancy-menu
