+++ source/views/menus/components/fancy-menu.js
@@ -1,6 +1,7 @@
 // @flow
+
 import React from 'react'
-import {View} from 'react-native'
+import {View, StyleSheet} from 'react-native'
 import {connect} from 'react-redux'
 import {updateMenuFilters} from '../../../flux'
 import type {TopLevelViewPropsType} from '../../types'
@@ -11,19 +12,18 @@ import type {
   ProcessedMealType,
   MenuItemContainerType,
 } from '../types'
-import type {FilterType, PickerType} from '../../components/filter'
-import {applyFiltersToItem} from '../../components/filter'
-import {fastGetTrimmedText} from '../../../lib/html'
-import {findMeal} from '../lib/find-menu'
 import fromPairs from 'lodash/fromPairs'
-import filter from 'lodash/filter'
-import find from 'lodash/find'
-import map from 'lodash/map'
-import flatten from 'lodash/flatten'
+import size from 'lodash/size'
 import values from 'lodash/values'
-import uniq from 'lodash/uniq'
+import {ListSeparator, ListSectionHeader} from '../../components/list'
+import type {FilterType} from '../../components/filter'
+import {applyFiltersToItem} from '../../components/filter'
+import SimpleListView from '../../components/listview'
+import {NoticeView} from '../../components/notice'
 import {FilterMenuToolbar} from './filter-menu-toolbar'
-import {MenuListView} from './menu'
+import {FoodItemRow} from './food-item-row'
+import {chooseMeal} from '../lib/choose-meal'
+import {buildFilters} from '../lib/build-filters'
 
 type FancyMenuPropsType = TopLevelViewPropsType & {
   applyFilters: (filters: FilterType[], item: MenuItemType) => boolean,
@@ -35,6 +35,14 @@ type FancyMenuPropsType = TopLevelViewPropsType & {
   menuCorIcons: MasterCorIconMapType,
   onFiltersChange: (f: FilterType[]) => any,
 };
+
+const leftSideSpacing = 28
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: 'white',
+  },
+})
 
 class FancyMenuView extends React.Component {
   static defaultProps = {
@@ -50,93 +58,7 @@ class FancyMenuView extends React.Component {
     }
 
     const foodItemsArray = values(foodItems)
-    filters = this.buildFilters(foodItemsArray, menuCorIcons, meals, now)
-    this.props.onFiltersChange(filters)
-  }
-
-  buildFilters(
-    foodItems: MenuItemType[],
-    corIcons: MasterCorIconMapType,
-    meals: ProcessedMealType[],
-    now: momentT
-  ): FilterType[] {
-    // Format the items for the stations filter
-    const stations = flatten(meals.map(meal => meal.stations))
-    const stationLabels = uniq(stations.map(station => station.label))
-    const allStations = stationLabels.map(label => ({title: label}))
-
-    // Grab the labels of the COR icons
-    const allDietaryRestrictions = map(corIcons, cor => ({
-      title: cor.label,
-      image: cor.image ? {uri: cor.image} : null,
-      detail: cor.description ? fastGetTrimmedText(cor.description) : '',
-    }))
-
-    // Decide which meal will be selected by default
-    const mealOptions = meals.map(m => ({label: m.label}))
-    const selectedMeal = this.findMeal(meals, [], now)
-
-    // Check if there is at least one special in order to show the specials-only filter
-    const stationNames = selectedMeal.stations.map(s => s.label)
-    const shouldShowSpecials = filter(foodItems, item =>
-      item.special && stationNames.includes(item.station)).length >= 1
-
-    return [
-      {
-        type: 'toggle',
-        key: 'specials',
-        enabled: shouldShowSpecials,
-        spec: {
-          label: 'Only Show Specials',
-          caption: 'Allows you to either see only the "specials" for today, or everything the location has to offer (e.g., condiments.)',
-        },
-        apply: {
-          key: 'special',
-        },
-      },
-      {
-        type: 'picker',
-        key: 'meals',
-        enabled: true,
-        spec: {
-          title: "Today's Menus",
-          options: mealOptions,
-          selected: {label: selectedMeal.label},
-        },
-        apply: {
-          key: 'label',
-        },
-      },
-      {
-        type: 'list',
-        key: 'stations',
-        enabled: false,
-        spec: {
-          title: 'Stations',
-          options: allStations,
-          mode: 'OR',
-          selected: allStations,
-        },
-        apply: {
-          key: 'station',
-        },
-      },
-      {
-        type: 'list',
-        key: 'dietary-restrictions',
-        enabled: false,
-        spec: {
-          title: 'Dietary Restrictions',
-          showImages: true,
-          options: allDietaryRestrictions,
-          mode: 'AND',
-          selected: [],
-        },
-        apply: {
-          key: 'cor_icon',
-        },
-      },
-    ]
+    this.props.onFiltersChange(buildFilters(foodItemsArray, menuCorIcons, meals, now))
   }
 
   props: FancyMenuPropsType;
@@ -155,22 +77,29 @@ class FancyMenuView extends React.Component {
     })
   }
 
-  findMeal(meals: ProcessedMealType[], filters: FilterType[], now: momentT): ProcessedMealType {
-    const mealChooserFilter: ?PickerType = find(filters, f => f.type === 'picker' && f.spec.title === "Today's Menus")
-    let selectedMeal = meals[0]
+  renderSectionHeader = (sectionData: MenuItemType[], sectionName: string) => {
+    const {filters, now, meals} = this.props
+    const {stations} = chooseMeal(meals, filters, now)
+    const menu = stations.find(m => m.label === sectionName)
+    const note = menu ? menu.note : ''
 
-    if (mealChooserFilter && mealChooserFilter.spec.selected) {
-      let label = mealChooserFilter.spec.selected.label
-      selectedMeal = meals.find(meal => meal.label === label)
-    } else {
-      selectedMeal = findMeal(meals, now)
-    }
+    return (
+      <ListSectionHeader
+        title={sectionName}
+        subtitle={note}
+        spacing={{left: leftSideSpacing}}
+      />
+    )
+  }
 
-    if (!selectedMeal) {
-      selectedMeal = {label: '', stations: [], starttime: '0:00', endtime: '0:00'}
-    }
-
-    return selectedMeal
+  renderSeparator = (sectionId: string, rowId: string) => {
+    return (
+      <ListSeparator
+        spacing={{left: leftSideSpacing}}
+        key={`${sectionId}-${rowId}`}
+        style={styles.separator}
+      />
+    )
   }
 
   render() {
@@ -182,8 +111,7 @@ class FancyMenuView extends React.Component {
       meals,
     } = this.props
 
-    const {label: mealName, stations: stationMenus} = this.findMeal(meals, filters, now)
-    const stationNotes = fromPairs(stationMenus.map(m => [m.label, m.note]))
+    const {label: mealName, stations: stationMenus} = chooseMeal(meals, filters, now)
 
     const filteredByMenu = stationMenus
       .map(menu => [
@@ -204,9 +132,22 @@ class FancyMenuView extends React.Component {
     const specialsFilterEnabled = Boolean(filters.find(f =>
       f.enabled && f.type === 'toggle' && f.spec.label === 'Only Show Specials'))
 
-    let message = ''
+    let messageView = null
     if (specialsFilterEnabled && stationMenus.length === 0) {
-      message = 'No items to show. There may be no specials today. Try changing the filters.'
+      messageView = (
+        <NoticeView
+          style={styles.container}
+          text='No items to show. There may be no specials today. Try changing the filters.'
+        />
+      )
+    }
+    if (!size(grouped)) {
+      messageView = (
+        <NoticeView
+          style={styles.container}
+          text='No items to show. Try changing the filters.'
+        />
+      )
     }
 
     return (
@@ -217,16 +158,27 @@ class FancyMenuView extends React.Component {
           filters={filters}
           onPress={this.openFilterView}
         />
-        <MenuListView
-          data={grouped}
-          stationNotes={stationNotes}
-          message={message}
-          corIcons={this.props.menuCorIcons}
-          // We can't conditionally show the star – wierd things happen, like
-          // the first two items having a star and none of the rest.
-          //badgeSpecials={!specialsFilterEnabled}
-          badgeSpecials
-        />
+        {messageView
+          ? messageView
+          : <SimpleListView
+            style={styles.container}
+            forceBottomInset={true}
+            data={grouped}
+            renderSeparator={this.renderSeparator}
+            renderSectionHeader={this.renderSectionHeader}
+          >
+            {(rowData: MenuItemType) =>
+              <FoodItemRow
+                data={rowData}
+                corIcons={this.props.menuCorIcons}
+                // We can't conditionally show the star – wierd things happen, like
+                // the first two items having a star and none of the rest.
+                //badgeSpecials={!specialsFilterEnabled}
+                badgeSpecials={true}
+                spacing={{left: leftSideSpacing}}
+              />}
+          </SimpleListView>
+        }
       </View>
     )
   }
