+++ source/views/menus/types.js
@@ -119,10 +119,6 @@ export type BonAppCafeInfoType = {
   }
 };
 
-export type ProcessedMenuPropsType = {
-  [key: string]: MenuItemType[],
-};
-
 export type ProcessedMealType = {|
   starttime: MilitaryTimeStringType,
   endtime: MilitaryTimeStringType,
